// packages/contracts/src/models/AdminCreatorPage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AdminCreatorOut } from './AdminCreatorOut';
export type AdminCreatorPage = {
    items: Array<AdminCreatorOut>;
    page: number;
    page_size: number;
    total: number;
};