// packages/contracts/src/models/AdminTransactionPage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AdminTransactionOut } from './AdminTransactionOut';
export type AdminTransactionPage = {
    items: Array<AdminTransactionOut>;
    page: number;
    page_size: number;
    total: number;
};