// packages/contracts/src/models/AdminUserDetailOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AdminUserDetailOut = {
    bio?: (string | null);
    country?: (string | null);
    created_at: string;
    discoverable?: boolean;
    display_name: string;
    email: string;
    featured?: boolean;
    handle?: (string | null);
    is_active: boolean;
    last_activity_at?: (string | null);
    last_login_at?: (string | null);
    last_login_ip?: (string | null);
    onboarding_state?: (string | null);
    phone?: (string | null);
    post_count?: number;
    role: string;
    signup_ip?: (string | null);
    subscriber_count?: number;
    total_earned_cents?: number;
    user_id: string;
    verified?: boolean;
};