// packages/contracts/src/models/AdminUserPage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AdminUserOut } from './AdminUserOut';
export type AdminUserPage = {
    items: Array<AdminUserOut>;
    page: number;
    page_size: number;
    total: number;
};