// packages/contracts/src/models/AdminUserPostPage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AdminUserPostOut } from './AdminUserPostOut';
export type AdminUserPostPage = {
    items: Array<AdminUserPostOut>;
    page: number;
    page_size: number;
    total: number;
};