// packages/contracts/src/models/AdminUserSubscriberPage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AdminUserSubscriberOut } from './AdminUserSubscriberOut';
export type AdminUserSubscriberPage = {
    items: Array<AdminUserSubscriberOut>;
    page: number;
    page_size: number;
    total: number;
};